feat(ossie): extract ossie-core as standalone in-tree module (#1396 L1)
Splits the Ossie / OSI parser layer out of saiku-service into its own
Maven module so downstream consumers can use it without pulling in
Mondrian, Spring, JAX-RS, or the rest of Saiku.

L1 of the library-extraction plan tracked in saiku#1396: internal
refactor only, no publication, no package rename. L2 will publish to
GitHub Packages; L3 splits to its own repo; L4 lands on Maven
Central.

## What moved

`saiku-core/saiku-service/src/main/java/org/saiku/service/schema/ossie/`
→ `saiku-core/ossie-core/src/main/java/org/saiku/service/schema/ossie/`

Files (13 classes + 4 tests):

- 11 DTOs (OssieDocument, SemanticModel, Dataset, Field, Metric,
  Relationship, Expression, DialectExpression, AiContext,
  CustomExtension, DimensionMeta)
- OssieYamlReader, OssieYamlWriter
- OssieYamlReaderTest (moved from saiku-sql, ported JUnit 4 → 5,
  package fixed)

## What stayed

- `MondrianToOssieConverter` (depends on Mondrian — belongs in
  saiku-service)
- `OssieYamlWriterTest` (uses MondrianToOssieConverter to build
  fixtures)
- Everything under `service/ossie/` (OssieDiscoverService,
  OssieQueryService, OssieShelfSqlTranslator, ai/) — these are
  Saiku service-layer concerns, not parser concerns

## Dependency graph after L1

    ossie-core   (jackson-databind + jackson-yaml only)
        ▲                          ▲
        │                          │
    saiku-service                saiku-sql (main)
        ▲                          ▲
        │                          │ (test)
        └──────── saiku-webapp ────┘

`saiku-sql` main no longer pulls in the full `saiku-service` graph;
it only needs the parser DTOs. Test scope still depends on
saiku-service for `SaikuOssieConnection`.

## Verification

- `mvn install -DskipTests`: BUILD SUCCESS across the reactor
- `mvn test` on ossie-core + saiku-service + saiku-sql:
  - ossie-core: 4/4 (reader tests)
  - saiku-service: 760/760 (0 failures, 1 pre-existing skip)
  - saiku-sql: SaikuOssieConnectionTest 2/2 + OssieShelfSqlTranslator
    13/13 + MondrianToOssieConverter 13/13 + OssieYamlWriter 4/4 +
    OssieAiValidator 12/12 + all Ossie AI paths clean
- OssieFuzzIT (6961 cases): `[Ossie combinatorial fuzz] passed 6912
  / 6912` + `[Ossie fuzz] passed 49 / 49`

Zero API break: all fully-qualified class names unchanged
(`org.saiku.service.schema.ossie.*`). External callers importing
these classes only see a different `.jar` on the classpath.

## Package rename?

Deliberately deferred to L2. Renaming (`org.saiku.service.schema.ossie`
→ `org.spiculedata.ossie` or `io.saiku.ossie`) is a coordinated
API-breaking change that makes sense when the artifact ships
externally — not for an internal reactor move.

Related: #1394, #1396.
