+++ saiku-core/ossie-core/pom.xml
@@ -1,0 +1,85 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.saikuanalytics</groupId>
+        <artifactId>saiku-core</artifactId>
+        <version>4.6.0</version>
+    </parent>
+
+    <artifactId>ossie-core</artifactId>
+    <packaging>jar</packaging>
+    <name>ossie - core (Ossie / OSI DTOs + YAML/JSON reader-writer)</name>
+    <description>Zero-dependency-beyond-Jackson library for reading, writing, and validating
+        Apache Ossie (Open Semantic Interchange) semantic model documents.
+
+        Provides:
+        - Immutable-ish DTO tree (SemanticModel, Dataset, Field, Metric, Relationship,
+          Expression, AiContext, CustomExtension, OssieDocument) modelling the OSI
+          v0.1.x / 0.2.x wire format.
+        - OssieYamlReader — reads YAML *or* JSON (Jackson YAMLMapper accepts both) into
+          the DTO tree. Consumes dbt Core 1.12's target/osi_document.json directly.
+        - OssieYamlWriter — writes a JSON-Schema-validated Ossie YAML back out;
+          bundled osi-schema.json validates against apache/ossie 0.2.0.dev0.
+
+        Intended to be independently consumable by any JVM tool that needs to read
+        or emit Ossie semantic models — Saiku uses it, but it's designed to have
+        zero coupling to Saiku, Spring, Calcite, or Mondrian.
+
+        Tracking issue: saiku#1396 (library extraction). L1 = in-tree module; L2+
+        publishes to Maven Central / OSSRH.</description>
+
+    <properties>
+        <maven.compiler.release>21</maven.compiler.release>
+    </properties>
+
+    <dependencies>
+        <!-- Jackson core + databind — DTO serialisation / deserialisation. -->
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+        </dependency>
+
+        <!-- YAML backend for Jackson — reads OSI YAML AND OSI JSON
+             (YAML is a JSON superset). -->
+        <dependency>
+            <groupId>com.fasterxml.jackson.dataformat</groupId>
+            <artifactId>jackson-dataformat-yaml</artifactId>
+        </dependency>
+
+        <!-- test-scoped: JSON Schema validator used by OssieYamlWriterTest to
+             confirm emitted YAML matches the bundled apache/ossie JSON schema.
+             Not needed at runtime — Jackson does the serialisation. -->
+        <dependency>
+            <groupId>com.networknt</groupId>
+            <artifactId>json-schema-validator</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <!-- Root pom pins compiler-plugin 3.1 + source/target 1.6 which
+                     doesn't understand the modern <release> flag. Pin 3.13.0
+                     locally (matches saiku-sql / saiku-service) so Java 21
+                     features compile cleanly. Don't 'fix' the root pom — other
+                     modules deliberately stay on the old surface. -->
+                <version>3.13.0</version>
+                <configuration>
+                    <release>${maven.compiler.release}</release>
+                    <encoding>UTF-8</encoding>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
