+++ saiku-core/ossie-core/src/main/java/org/saiku/service/schema/ossie/model/CustomExtension.java
