+++ saiku-core/pom.xml
@@ -12,6 +12,7 @@
     <version>4.6.0</version>
 
     <modules>
+        <module>ossie-core</module>
         <module>saiku-olap-util</module>
         <module>saiku-semantic</module>
         <module>saiku-service</module>
