+++ saiku-core/saiku-sql/pom.xml
@@ -22,12 +22,21 @@
     </properties>
 
     <dependencies>
-        <!-- Ossie model + YAML round-trip live in saiku-service; the SchemaFactory
-             reads the same POJO tree the Mondrian->Ossie exporter produces. -->
+        <!-- Ossie DTOs + YAML/JSON reader/writer, extracted to ossie-core so
+             saiku-sql only depends on the parser layer, not on the full
+             saiku-service (Mondrian, Spring, JAX-RS). Tests still need
+             saiku-service for the SaikuOssieConnection wrapper — pulled in
+             below with test scope. Tracking issue: saiku#1396 L1. -->
+        <dependency>
+            <groupId>org.saikuanalytics</groupId>
+            <artifactId>ossie-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <dependency>
             <groupId>org.saikuanalytics</groupId>
             <artifactId>saiku-service</artifactId>
             <version>${project.version}</version>
+            <scope>test</scope>
         </dependency>
 
         <!-- Apache Calcite - the SQL parser, planner, and virtual-schema adapter
